added ability to regiister new user with customer data, CustomerData DateOfBirth changed to DateTime, AdminPanel UsersList can Create, Edit and Delete users, Email in User and Customer Data edits is not editable (hidden / readonly), in Project WebMVC added new folder "Data", folder data has now datafiles for Users, CustomerData, AdditionalServices and PaymentHistory, added new model AdditionalService, added added Payment History in Customer Data Menu

+++ HotelHero/ReservationsDatabase/Reservation.cs
@@ -17,11 +17,11 @@ namespace HotelHero.ReservationsDatabase
         public DateTime CheckInDate { get; }
         public DateTime CheckOutDate { get; }
         public int AmountOfPeople { get; }
-        public int CostPerNight { get; }
+        public decimal CostPerNight { get; }
         public ReservationStatus Status { get; set; }
         public string ReservationUser { get; set; }
 
-        public Reservation(int id, Hotel hotel, DateTime checkInDate, DateTime checkOutDate, int amountOfPeople, int costPerNight, ReservationStatus reservationStatus, string reservationUser)
+        public Reservation(int id, Hotel hotel, DateTime checkInDate, DateTime checkOutDate, int amountOfPeople, decimal costPerNight, ReservationStatus reservationStatus, string reservationUser)
         {
             Id = id;
             Hotel = hotel;
